+++ src/main/java/com/maxim/pojo/kbar/StockBar.java
@@ -1,4 +1,5 @@
 package com.maxim.pojo.kbar;
+import com.alibaba.fastjson2.annotation.JSONField;
 import java.time.*;
 
 
@@ -8,7 +9,7 @@ public class StockBar extends KBar{
     }
 
     public StockBar(String symbol, LocalDate tradeDate, LocalTime tradeTime,
-                    Double open, Double high, Double low, Double close, Double volume){
+                    Double open, Double high, Double low, Double close, Integer volume){
         super(symbol, tradeDate, tradeTime, open, high, low, close, volume);
     }
 }
