+++ src/main/java/com/maxim/pojo/summary/FutureSummary.java
@@ -7,7 +7,7 @@ public class FutureSummary extends Summary{
     public Double profit;
     public Double realTimePrice; // 最新价格
     public Double realTimeProfit; // 最新浮盈浮亏
-    public FutureSummary(Double ori_price, Double total_vol, Double static_profit, Double static_loss, Double dynamic_profit,
+    public FutureSummary(Double ori_price, Integer total_vol, Double static_profit, Double static_loss, Double dynamic_profit,
                          Double dynamic_loss){
         super(ori_price, total_vol, static_profit, static_loss, dynamic_profit, dynamic_loss);
         this.dynamic_monitor = 0;
@@ -17,7 +17,7 @@ public class FutureSummary extends Summary{
     }
 
     // 开仓时触发回调
-    public void openUpdate(Double price, Double vol, Double static_profit, Double static_loss, Double dynamic_profit, Double dynamic_loss){
+    public void openUpdate(Double price, Integer vol, Double static_profit, Double static_loss, Double dynamic_profit, Double dynamic_loss){
         // 更新基本属性
         this.static_profit = static_profit;
         this.static_loss = static_loss;
@@ -26,9 +26,9 @@ public class FutureSummary extends Summary{
 
         // 更新vol & ori_price(持仓买入均价)
         Double ori_price = this.ori_price;
-        Double vol0 = this.total_vol;
+        Integer vol0 = this.total_vol;
         double amount0 = ori_price * total_vol;
-        Double vol1 = vol0 + vol;
+        Integer vol1 = vol0 + vol;
         Double amount1 = amount0 + price * vol;
 
         // 赋值回sum
@@ -36,7 +36,7 @@ public class FutureSummary extends Summary{
         this.ori_price = amount1 / vol1;  // 更新持仓均价
     }
 
-    public void closeLongUpdate(Double price, Double vol){
+    public void closeLongUpdate(Double price, Integer vol){
         /*
         更新基本属性
         对于多头仓位视图, 以price卖出vol后, summary的变动
@@ -51,7 +51,7 @@ public class FutureSummary extends Summary{
         this.realTimeProfit -= profit;
     }
 
-    public void closeShortUpdate(Double price, Double vol){
+    public void closeShortUpdate(Double price, Integer vol){
         /*
         更新基本属性
         对于空头仓位视图, 以price卖出vol后, summary的变动
